+++ packages/cms-rag/client/index.ts
@@ -1,6 +1,7 @@
 export * from "./api.js";
 export { MarkdownEditor } from "./MarkdownEditor.js";
 export { MediaPicker } from "./MediaPicker.js";
+export { FilePicker } from "./FilePicker.js";
 export { MediaLibrary } from "./MediaLibrary.js";
 export { EntityPanel } from "./EntityPanel.js";
 export { ContentManager, type CustomPanel } from "./ContentManager.js";